refactor(editor): Group module SDK into types/ and registries/ dirs (no-changelog)
Address Phase A review feedback:
- Move the package sources under src/types/ and src/registries/ (was flat), so
  types and registries each live in their own directory.
- Rename CommandBarContribution -> CommandBarEntry, aligning with the existing
  CommandBarItem vocabulary in the command bar.

The public entry (src/index.ts) keeps the same surface — it re-exports the same
types and registries through the new barrels, so consumers are unaffected.

Linear: https://linear.app/n8n/issue/CAT-3685
