+++ packages/frontend/@n8n/frontend-module-sdk/src/registries/commandRegistry.ts
@@ -1,9 +1,9 @@
-import type { CommandBarContribution } from './command';
+import type { CommandBarEntry } from '../types/command';
 
-const commands = new Map<string, CommandBarContribution>();
-const listeners = new Set<(commands: CommandBarContribution[]) => void>();
+const commands = new Map<string, CommandBarEntry>();
+const listeners = new Set<(commands: CommandBarEntry[]) => void>();
 
-export function getAll(): CommandBarContribution[] {
+export function getAll(): CommandBarEntry[] {
 	return Array.from(commands.values());
 }
 
@@ -12,7 +12,7 @@ function notifyListeners(): void {
 	listeners.forEach((listener) => listener(snapshot));
 }
 
-export function register(command: CommandBarContribution): void {
+export function register(command: CommandBarEntry): void {
 	if (commands.has(command.id)) {
 		console.warn(`Command with id "${command.id}" is already registered. Skipping.`);
 		return;
@@ -27,7 +27,7 @@ export function unregister(id: string): void {
 	}
 }
 
-export function get(id: string): CommandBarContribution | undefined {
+export function get(id: string): CommandBarEntry | undefined {
 	return commands.get(id);
 }
 
@@ -35,7 +35,7 @@ export function has(id: string): boolean {
 	return commands.has(id);
 }
 
-export function subscribe(listener: (commands: CommandBarContribution[]) => void): () => void {
+export function subscribe(listener: (commands: CommandBarEntry[]) => void): () => void {
 	listeners.add(listener);
 	return () => {
 		listeners.delete(listener);
