+++ packages/frontend/@n8n/frontend-module-sdk/src/registries/modalRegistry.ts
@@ -1,4 +1,4 @@
-import type { ModalDefinition } from './module.types';
+import type { ModalDefinition } from '../types/modal';
 
 const modals = new Map<string, ModalDefinition>();
 const listeners = new Set<(modals: Map<string, ModalDefinition>) => void>();
