+++ packages/frontend/@n8n/frontend-module-sdk/src/types/command.ts
@@ -3,14 +3,14 @@
  * stable subset of the design-system `CommandBarItem`) so the module contract
  * does not couple to command-bar internals.
  */
-export interface CommandBarContribution {
+export interface CommandBarEntry {
 	id: string;
 	title: string;
 	section?: string;
 	keywords?: string[];
 	icon?: string;
 	handler?: () => void | Promise<void>;
-	children?: CommandBarContribution[];
+	children?: CommandBarEntry[];
 	/**
 	 * Route names this command is active on. Undefined/empty means all views.
 	 */
