+++ packages/frontend/@n8n/frontend-module-sdk/src/types/descriptor.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect } from 'vitest';
 
-import type { FrontendModuleDescription } from './module.types';
+import type { FrontendModuleDescription } from './descriptor';
 
 describe('FrontendModuleDescription', () => {
 	it('accepts a v1-shaped descriptor (no v2 fields)', () => {
