+++ packages/frontend/@n8n/frontend-module-sdk/src/types/locale.ts
