+++ packages/frontend/@n8n/frontend-module-sdk/src/types/resource.ts
@@ -1,0 +1,9 @@
+/**
+ * Metadata a module contributes to describe a resource type it owns. Consumed by
+ * generic list surfaces (e.g. `ResourcesListLayout`).
+ */
+export type ResourceMetadata = {
+	key: string;
+	displayName: string;
+	i18nKeys?: Record<string, string>;
+};
