+++ packages/frontend/@n8n/frontend-module-sdk/src/types/shortcut.ts
